+++ reusables/default_list.py
@@ -3,13 +3,13 @@
 
 
 class defaultlist(list):
-    def __init__(self, *args, factory=None, **kwargs):
+    def __init__(self, *args, factory=lambda: None, **kwargs):
         super().__init__(*args, **kwargs)
+        if not callable(factory):
+            raise Exception("The factory must be callable")
         self.factory = factory
 
     def __getitem__(self, index):
         if index >= len(self):
-            diff = index - len(self) + 1
-            for i in range(diff):
-                self.append(self.factory())
+            self.extend(self.factory() for _ in range(index - len(self) + 1))
         return super().__getitem__(index)