+++ test/test_wrappers.py
@@ -3,7 +3,7 @@
 import time
 import os
 
-from .common_test_data import *
+from .common_test_data import BaseTestClass
 
 from reusables import (
     unique,
@@ -131,7 +131,7 @@ class TestWrappers(BaseTestClass):
         correctly in the logs.
         """
         setup_logger("my_logger", file_path="out.log")
-        message = "I would like to take this moment to say something " "interesting has happened. "
+        message = "I would like to take this moment to say something interesting has happened. "
 
         @log_exception("my_logger", message=message)
         def unique_function_5():
@@ -185,7 +185,3 @@ class TestWrappers(BaseTestClass):
 
         b()
         assert ouch() == 10
-
-
-if __name__ == "__main__":
-    unittest.main()