+++ CHANGES.rst
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+Version 1.0.0
+-------------
+
+- Fixing warning about escape sequences
+- Fixing default of defaultlist
+- Changing testing from Travis CI to github actions
+- Removing Python 2.x support
+- Removing scandir options
+- Removing tests for web
+
 Version 0.9.6
 -------------